R6 class representing queryBuilderConfig object.

R6 class representing queryBuilderConfig object.

R6 class representing queryBuilderConfig object.

Details

The object is responsible for storing definitions for operators and conditions that are used to generate query expression. It also allows to manage its objects by the provided methods.

Methods

Public methods

queryBuilderConfigClass$new()

queryBuilderConfigClass$add()

queryBuilderConfigClass$remove()

queryBuilderConfigClass$get_from_private()

queryBuilderConfigClass$set_to_private()

queryBuilderConfigClass$reset()

queryBuilderConfigClass$clone()

Method new()

Create queryBuilderConfig object with initialized conditions and operators.

Usage

queryBuilderConfigClass$new(
  conditions = default_conditions,
  operators = default_operators,
  ...
)

Arguments

conditions

Conditions.

operators

Operators.

...

Unused.

Returns

The object of class `queryBuilderConfig`.

Method add()

Add conditions and conditions to queryBuilderConfig object.

Usage

queryBuilderConfigClass$add(conditions = NULL, operators = NULL)

Arguments

conditions

Conditions.

operators

Operators.

Method remove()

Remove conditions or operators from queryBuilderConfig object.

Usage

queryBuilderConfigClass$remove(conditions_id = NULL, operators_id = NULL)

Arguments

conditions_id

Id of conditions to remove.

operators_id

Id of operators to remove.

Method get_from_private()

Get private elements from queryBuilderConfig object.

Usage

queryBuilderConfigClass$get_from_private(name)

Arguments

name

Name of the element to get.

Method set_to_private()

Set private elements to queryBuilderConfig object.

Usage

queryBuilderConfigClass$set_to_private(name, value)

Arguments

name

Name of the element to set.

value

New element value.

Method reset()

Restore default conditions and conditions of queryBuilderConfig object and clear out remaining pprivate objects.

Usage

queryBuilderConfigClass$reset()

Method clone()

The objects of this class are cloneable with this method.

Usage

queryBuilderConfigClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.